obviously you can't use sealed class on java8

+++ common/src/main/java/dev/latvian/kubejs/block/MapColorHelper.java
@@ -128,7 +128,7 @@ public record MapColorHelper(int id, String name, MaterialColor color, Vector3f 
 
 				if (dist < lastDist) {
 					closest = helper;
-					lastDist = (float) dist;
+					lastDist = dist;
 				}
 			}
 		}
